Fix some more integer conversion warnings

+++ Mac-Old/Mac-Old/Controllers/JVTranscriptCriterionController.h
@@ -8,7 +8,7 @@ typedef enum JVTranscriptCriterionFormat { // cooresponds to the nib tab view id
 } JVTranscriptCriterionFormat;
 
 /// corresponds to the nib menu tags
-typedef NS_ENUM(int, JVTranscriptCriterionKind) {
+typedef NS_ENUM(NSInteger, JVTranscriptCriterionKind) {
 	JVTranscriptMessageBodyCriterionKind = 1,
 	JVTranscriptSenderNameCriterionKind,
 	JVTranscriptDateReceivedCriterionKind,
@@ -36,7 +36,7 @@ typedef NS_ENUM(int, JVTranscriptCriterionKind) {
 };
 
 /// corresponds to the nib menu tags
-typedef NS_ENUM(int, JVTranscriptCriterionOperation) {
+typedef NS_ENUM(NSInteger, JVTranscriptCriterionOperation) {
 	JVTranscriptNoCriterionOperation = 0,
 	JVTranscriptTextMatchCriterionOperation = 1,
 	JVTranscriptTextDoesNotMatchCriterionOperation,
@@ -51,7 +51,7 @@ typedef NS_ENUM(int, JVTranscriptCriterionOperation) {
 };
 
 /// corresponds to the nib menu tags
-typedef NS_ENUM(int, JVTranscriptCriterionQueryUnits) {
+typedef NS_ENUM(NSInteger, JVTranscriptCriterionQueryUnits) {
 	JVTranscriptNoCriterionQueryUnits = 0,
 	JVTranscriptSecondCriterionQueryUnits = 1,
 	JVTranscriptMinuteCriterionQueryUnits,
